Moving application/window discovery into dialog
This change simplifies the Nautilus application/window discovery and moves it
into the dialog itself. This also simplifies the constructor of each dialog by
moving the shared child into the parent class.

By moving this functionality the function it utils is no longer needed. The
dialog was the only place this was used.