Adding new utility method and Init DB menu
This update adds a new utility function that will transform the Nautilus
FileInfo object into a string path. This is used through the code to simplify
value checks.

This update also adds a new Init DB menu item for TMSU. This allows a TMSU
database to be created in folder locations only.