Updating to rewrite how commands are run
Moved the queue runner into its own class with a singleton. This allows for
the queue to be shared among the whole plugin easily. I also found that if a
timeout is added, then Nautilus will pay more attention to the plugin. This is
also how the speed of commands is controlled in the queue as a by product.

The commands have all been completely rewritten into their own classes that
extend from one another and provide mixins for common functionality. This
should support further development with more commands in the future.