Fixing tag listing error
When all tags are listed, there is no prefix to the tag list. However, when
tags are listed for a file there is a prefix. This update fixes that by only
removing the prefix when it should exist. This was previously removing the
first tag of the all tags list.