Updating tmsu db check for update_file_info_full
This changes how update_file_info_full works by checking the parent directory
for a TMSU database. This method is only called with children of the directory
we are in. Adding a simple check against the parent of the first file then
prevents checks against other children if we're not in a TMSU db.

Logging was also updated to shift some information out further.